+++ src/Support/Tests/Dummies/ConfigureMe.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the ConfigureMe class.
+ *
+ * @copyright   Copyright (c) 2023 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-01-10
+ *
+ */
+
+namespace Vanilo\Support\Tests\Dummies;
+
+use Illuminate\Database\Eloquent\Model;
+
+class ConfigureMe extends Model
+{
+}